Music library UI change
